requirements.txt
youtube_transcript_api==0.4.1
streamlit==0.87.0
google_generativeai==0.1.0
python-dotenv==0.19.0
pathlib==1.0.1